Change Required in SourceDBInstanceIdentifier
While creating RDS read replica in the same region in a different subnet, If we enter the ID of the source DB instance for the SourceDBInstanceIdentifier property as per the document it leads to the following error :"DbSubnetGroupName should not be specified for read replicas that are created in the same region as the master". When the ID is replaced with ARN of the DB instance it gets replicated successfully.